TYP: Sane defaults for the  platform-specific ``NBitBase`` types.
This will help for those that don't use the mypy plugin.
