+++ numpy/_typing/_nbit.py
@@ -1,17 +1,19 @@
 """A module with the precisions of platform-specific `~numpy.number`s."""
 
-from typing import Any
+from typing import TypeAlias
+from ._nbit_base import _8Bit, _16Bit, _32Bit, _64Bit, _96Bit, _128Bit
+
 
 # To-be replaced with a `npt.NBitBase` subclass by numpy's mypy plugin
-_NBitByte = Any
-_NBitShort = Any
-_NBitIntC = Any
-_NBitIntP = Any
-_NBitInt = Any
-_NBitLong = Any
-_NBitLongLong = Any
+_NBitByte: TypeAlias = _8Bit
+_NBitShort: TypeAlias = _16Bit
+_NBitIntC: TypeAlias = _32Bit
+_NBitIntP: TypeAlias = _32Bit | _64Bit
+_NBitInt: TypeAlias = _NBitIntP
+_NBitLong: TypeAlias = _32Bit | _64Bit
+_NBitLongLong: TypeAlias = _64Bit
 
-_NBitHalf = Any
-_NBitSingle = Any
-_NBitDouble = Any
-_NBitLongDouble = Any
+_NBitHalf: TypeAlias = _16Bit
+_NBitSingle: TypeAlias = _32Bit
+_NBitDouble: TypeAlias = _64Bit
+_NBitLongDouble: TypeAlias = _64Bit | _96Bit | _128Bit
